--- markup/components/card/card.js
@@ -1,3 +1,0 @@
-export default function () {
-    return (<h1>Card</h1>);
-}
