+++ markup/components/events/data/data.js
@@ -1,4 +1,4 @@
-let dataEvents = {events: {
+let dataMain = {events: {
     "events": [
         {
             "type": "info",
