+++ markup/components/events/events.pug
@@ -1,11 +1,10 @@
 include ../card/card.pug
 
 // data - array of events (cards)
-mixin main(data)
+mixin events(data)
     main.events
         div.events__title
             h1 Лента событий
-        //#main.test
         .card__container
             - let events = data;
 
@@ -13,6 +12,3 @@ mixin main(data)
                 +card(val)
             else 
                 div There are no values
-
-
-
