+++ markup/components/events/events.scss
@@ -54,4 +54,3 @@
     }
 
 }
-
