+++ markup/components/header/data/data.js
@@ -1,0 +1,26 @@
+let dataHeader = {header: {
+    defaults: {
+        menuItems: [
+            {
+                title: 'События',
+                url: '/events.html'
+            },
+            {
+                title: 'Cводка',
+                url: '#'
+            },
+            {
+                title: 'Устройство',
+                url: '#'
+            },
+            {
+                title: 'Сценарии',
+                url: '#'
+            },
+            {
+                title: 'Видео наблюдение',
+                url: '/cctv.html'
+            }
+        ]
+    }
+}}
