--- markup/components/header/header.js
@@ -1,1 +1,0 @@
-
