+++ markup/components/header/header.pug
@@ -3,5 +3,5 @@ include ../logo/logo.pug
 
 mixin header(data)
     header.header
-        +logo.header__item()
-        +menu.header__item()
+        +logo().header__item
+        +menu(data.menuItems).header__item
