--- markup/components/logo/logo.js
@@ -1,1 +1,0 @@
-
