+++ markup/components/logo/logo.pug
@@ -1,3 +1,4 @@
 mixin logo()
-    .logo.icon
-        .icon__logo
+    a(href='/index.html')    
+        .logo.icon
+            .icon__logo
