+++ markup/components/logo/logo.scss
@@ -1,12 +1,12 @@
-.icon__logo{
+.icon__logo {
     //@include bg-svg($Яндекс-Дом);
 
     grid-area: "logo";
 
     @include bg($dom-logo);
-  
+
     @media screen and (max-width: 375px) {
-      margin-left: 20px;
+        margin-left: 20px;
     }
-  }
-  
+}
+
