+++ markup/components/main/main.pug
@@ -1,12 +1,2 @@
 mixin main(data)
-    main.events
-        div.events__title
-            h1 Лента событий
-        //#main.test
-        .card__container
-            - let events = data;
-
-            each val in events
-                +card(val)
-            else 
-                div There are no values
+    main.main
