+++ markup/pages/cctv.pug
@@ -9,7 +9,7 @@ html(class="no-js", lang="ru")
     body.page
         |%=symbols=%
 
-        +header()
+        +header(header.defaults)
 
         +footer()
